fix: four wave tests that only failed where the environment differed
The 3.8-fallback test for assert_in_loop read ast.unparse before deleting it, so
it raised on 3.8 - the very interpreter whose fallback it exists to cover.
showcase_df_columns still printed with use_print=False because an old
'or not HAS_IPYTHON' clause overrode the explicit flag, so that finding was only
closed where IPython happens to be installed. n_unique on a Decimal column is
unsupported by the polars that installs on 3.8, so the count is now taken after
a value-preserving cast. Four tests needed flask or polars_ds, neither of which
CI installs, and now say so.

The dashlib tests moved to their own module: their new precondition pushed a
file already sitting at the 1000-LOC budget over it, and they share that one
precondition.

Also carries a downstream scan report - the freshly-fixed scanners run over
py-ci-shared and mlframe, with per-check false-positive rates measured.

+++ src/pyutilz/data/polarslib/binning.py
@@ -267,7 +267,9 @@ def bin_numerical_columns(
         skipped_clips = []
         if min_nuniques_to_clip:
             # do not apply clipping if # of unique values is too low (under 10)
-            n_uniques_dict = df.lazy().select(pl.col(clips.keys()).n_unique()).collect().row(0, named=True)
+            # Cast before counting: older polars refuses n_unique on Decimal, and the count is
+            # identical either way because the cast is value-preserving for the widths used here.
+            n_uniques_dict = df.lazy().select(pl.col(clips.keys()).cast(pl.Float64).n_unique()).collect().row(0, named=True)
             for col, nuniques in n_uniques_dict.items():
                 if nuniques < min_nuniques_to_clip:
                     for field in "min max".split():
